HackathonOrchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
openai>=1.0.0

# Discord Bot Dependencies
discord.py==2.3.2
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Database and Vector Storage
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI and ML
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Security and Authentication
PyJWT==2.8.0
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Rate Limiting and Caching
redis==5.0.1
slowapi==0.1.9

# Validation and Serialization
pydantic==2.5.2
marshmallow==3.20.2
